+++ tests/e2e/core-tests/specs/index.js
@@ -21,6 +21,7 @@ const runSingleProductPageTest = require( './shopper/front-end-single-product.te
 const runVariableProductUpdateTest = require( './shopper/front-end-variable-product-updates.test' );
 const runCheckoutCreateAccountTest = require( './shopper/front-end-checkout-create-account.test' );
 const runCheckoutLoginAccountTest = require( './shopper/front-end-checkout-login-account.test' );
+const runCartRedirectionTest = require( './shopper/front-end-cart-redirection.test' );
 
 // Merchant tests
 const runAddNewShippingZoneTest = require ( './merchant/wp-admin-settings-shipping-zones.test' );
@@ -69,6 +70,7 @@ const runShopperTests = () => {
 	runVariableProductUpdateTest();
 	runCheckoutCreateAccountTest();
 	runCheckoutLoginAccountTest();
+  runCartRedirectionTest();
 };
 
 const runMerchantTests = () => {
@@ -144,5 +146,6 @@ module.exports = {
 	runCheckoutCreateAccountTest,
   runImportProductsTest,
 	runCheckoutLoginAccountTest,
-	runMyAccountCreateAccountTest
+	runCartRedirectionTest,
+	runMyAccountCreateAccountTest,
 };